Move test_positive_update_capsule_with_hammer to SatelliteMaintain/Rocket (#21790)
The test invokes 'satellite-maintain' via REX job template 'Capsule Update
Playbook', making it a maintain operation, not a ForemanProxy concern. Moving
it to tests/foreman/maintain/test_upgrade.py gives it the correct
CaseComponent (SatelliteMaintain) and Team (Rocket) ownership inherited from
the module header, fixing miscategorised failures in foremanproxy 6.19.2
test results.

Removes the 'settings' import from test_capsule.py that was added solely for
this test in PR #19933.

:verifies: SAT-46508

(cherry picked from commit 071222a9e5463ab9681faee007c491732551a57f)
